STIX2 import Refactor
